url-scraper-blob/requirements.txt
azure-functions
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
azure-core==1.31.0
azure-functions==1.21.1
azure-storage-blob==12.23.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
cssselect==1.2.0
distro==1.9.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
isodate==0.6.1
jiter==0.5.0
lxml==5.3.0
lxml_html_clean==0.2.2
numpy==2.1.1
openai==1.50.2
pandas==2.2.3
parse==1.20.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.1.1
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-html==0.10.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
w3lib==2.2.1
websockets==10.4
zipp==3.20.2